Zombie Follow Changes
Map initializing and moving are in separate scripts. (This is to make so each navigation agent does not have to create their own map).
Added script to change path mode to diagonal when the agent has a clear view of the target, and is vertical when walls are in the way. (This is so that agents do not walk through walls but still walk relatively straight when there are no walls in the way.
Added changing targets script to change targets when a closer target is available.
Not sure how efficient these scripts are but I hope to use them in my multiplayer zombie game.